fixed login , added Landing page , implemented logout
